+++ lib/proto_pharm/stub_components/failure_response.rb
@@ -1,15 +1,13 @@
 # frozen_string_literal: true
 
-require "active_support/core_ext/object/blank"
-
 module ProtoPharm
   module StubComponents
     module FailureResponse
       # @param code [String, Symbol] A gRPC failure code, such as not_found or invalid_argument. Default: :invalid_argument
       # @param message [String] A message to pass back with the exception
-      # @param metadata [Hash] A hash of metadata to be passed back with the exception
-      def to_fail_with(code = :invalid_argument, message = "unknown cause", metadata: {})
-        to_raise(exception_class(code).new(message, metadata))
+      # @param metadata [Hash] A hash of metadata to be passed back with the exception. See {ProtoPharm::MetadataSerializers::Base.serialize}.
+      def to_fail_with(code = :invalid_argument, message = "unknown cause", **metadata)
+        to_raise(exception_class(code).new(message, metadata_serializer.serialize(code: code, message: message, **metadata)))
       end
 
       def to_fail
@@ -17,6 +15,10 @@ module ProtoPharm
       end
 
       private
+
+      def metadata_serializer
+        Configuration.config.metadata_serializer
+      end
 
       def exception_class(code)
         class_name = code.to_s.camelize