+++ spec/spec_helper.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
-require 'bundler/setup'
-require 'pry'
-require 'faker'
-require 'rspice'
+require "bundler/setup"
+require "pry"
+require "faker"
+require "rspice"
 
 require "proto_pharm"
 require "proto_pharm/rspec"
+require "directive/spec_helper"
 
 require_relative "examples/hello/hello_services_pb"
 require_relative "examples/request/request_services_pb"
@@ -14,7 +15,7 @@ require_relative "examples/hello/hello_client"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = '.rspec_status'
+  config.example_status_persistence_file_path = ".rspec_status"
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!