+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-# 0.6.0
+## Upcoming
+
+## 0.7.0
+*Release Date: 2021-02-19*
+
+### Additions
+- Add configurable metadata serializer ([#6](https://github.com/Freshly/proto_pharm/pull/6))
+- Add metadata serializer for Gruf ([#7](https://github.com/Freshly/proto_pharm/pull/7))
+- Stub responses using a block ([#8](https://github.com/Freshly/proto_pharm/pull/8))
+- Net connect method cleanup ([#9](https://github.com/Freshly/proto_pharm/pull/9))
+
+## 0.6.0
 *Release Date: 2020-12-07*
 
 - Rename gem to `proto_pharm`